Separate lists view test files
